8655: Move verifyUserPendingEmailLamnda.ts to /lambdas/public-api
