+++ web-api/src/app-public.ts
@@ -91,7 +91,7 @@ import { orderPublicSearchLambda } from './lambdas/public-api/orderPublicSearchL
 import { todaysOpinionsLambda } from './lambdas/public-api/todaysOpinionsLambda';
 import { todaysOrdersLambda } from './lambdas/public-api/todaysOrdersLambda';
 import { getDbReader } from '@web-api/database';
-import { verifyUserPendingEmailLambda } from './lambdas/users/verifyUserPendingEmailLambda';
+import { verifyUserPendingEmailLambda } from './lambdas/public-api/verifyUserPendingEmailLambda';
 
 /** Case */
 {
@@ -181,5 +181,8 @@ app.get('/public-api/judges', lambdaWrapper(getPublicJudgesLambda));
 }
 
 {
-  app.put('/public-api/verify-email', lambdaWrapper(verifyUserPendingEmailLambda));
+  app.put(
+    '/public-api/verify-email',
+    lambdaWrapper(verifyUserPendingEmailLambda),
+  );
 }
