+++ web-api/src/lambdas/public-api/verifyUserPendingEmailLambda.ts
